More changes for pylint compliance
