+++ flaskr/account/templates/edit_account.html
@@ -19,6 +19,9 @@
         <div class="form-group">
             <input id='email' class="form-control" name="email" value={{account.email}} type="text">
         </div>
+        <div class="form-group">
+            <input id='language' class="form-control" name="email" value={{account.language}} type="text">
+        </div>
 
         <button id='update' class="btn" type="submit" name="submit" value="update">{{_('Update')}}</button>
         <button id='cancel' class="btn btn-primary" type="submit" name="submit" value="cancel">{{_('Cancel')}}</button>
